Fix string corruption in vtable extraction
Save the result of getName().str() in a local std::string variable.
Previously c_str() was called on a temporary that was destructed at the
end of the expression, resulting on a dangling pointer.

This could lead to corrupted strings being written to stoat_vtable.txt,
which stoat then failed to parse.